feat: keyboard input hook
-Add a custom hook to manage all of the keyboard inputs for typing test textbox logic.
